+++ vector/src/main/java/im/vector/riotx/features/home/room/breadcrumbs/BreadcrumbsController.kt
@@ -19,6 +19,7 @@ package im.vector.riotx.features.home.room.breadcrumbs
 import android.view.View
 import com.airbnb.epoxy.EpoxyController
 import im.vector.matrix.android.api.util.toMatrixItem
+import im.vector.riotx.core.epoxy.zeroItem
 import im.vector.riotx.core.utils.DebouncedClickListener
 import im.vector.riotx.features.home.AvatarRenderer
 import javax.inject.Inject
@@ -45,9 +46,13 @@ class BreadcrumbsController @Inject constructor(
     override fun buildModels() {
         val safeViewState = viewState ?: return
 
+        // Add a ZeroItem to avoid automatic scroll when the breadcrumbs are updated from another client
+        zeroItem {
+            id("top")
+        }
+
         // An empty breadcrumbs list can only be temporary because when entering in a room,
         // this one is added to the breadcrumbs
-
         safeViewState.asyncBreadcrumbs.invoke()
                 ?.forEach {
                     breadcrumbsItem {